Keep rspec.sha guard self-contained
