+++ tools/check-rspec-sha.ts
@@ -14,15 +14,48 @@
  * You should have received a copy of the Sonar Source-Available License
  * along with this program; if not, see https://sonarsource.com/license/ssal/
  */
-import { dirname, join } from 'node:path';
+import { existsSync, readFileSync } from 'node:fs';
+import { dirname, join, resolve } from 'node:path';
 import { fileURLToPath } from 'node:url';
-import { readRootRspecShaOverride } from '../packages/shared/src/helpers/rspec.js';
 
-const __dirname = dirname(fileURLToPath(import.meta.url));
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = dirname(__filename);
 const ROOT_DIR = join(__dirname, '..');
+const DEFAULT_ERROR_MESSAGE = 'Failed to evaluate root rspec.sha override.';
 
-try {
-  const pinnedSha = readRootRspecShaOverride(ROOT_DIR);
+export function readRootRspecShaOverride(projectRoot: string) {
+  const overrideFile = join(projectRoot, 'rspec.sha');
+  if (!existsSync(overrideFile)) {
+    return null;
+  }
+
+  const pinnedSha = readFileSync(overrideFile, 'utf-8').trim();
+  if (!pinnedSha) {
+    throw new Error('rspec.sha is empty');
+  }
+
+  return pinnedSha;
+}
+
+function getErrorMessage(error: unknown) {
+  if (typeof error === 'string') {
+    return error;
+  }
+
+  if (
+    typeof error === 'object' &&
+    error !== null &&
+    'message' in error &&
+    typeof error.message === 'string'
+  ) {
+    return error.message;
+  }
+
+  return DEFAULT_ERROR_MESSAGE;
+}
+
+export function run(projectRoot = ROOT_DIR) {
+  const pinnedSha = readRootRspecShaOverride(projectRoot);
 
   if (pinnedSha === null) {
     console.log('No root rspec.sha override found.');
@@ -32,7 +65,13 @@ try {
     );
     process.exitCode = 1;
   }
-} catch (error: any) {
-  console.error(error?.message ?? 'Failed to evaluate root rspec.sha override.');
-  process.exitCode = 1;
+}
+
+if (process.argv[1] !== undefined && resolve(process.argv[1]) === __filename) {
+  try {
+    run();
+  } catch (error: unknown) {
+    console.error(getErrorMessage(error));
+    process.exitCode = 1;
+  }
 }
